Extract `SwiftVersionProviding` from `System`
This enables caching without having a coupling to `System.shared`
