+++ Sources/TuistSupport/System/Systeming.swift
@@ -73,18 +73,6 @@ public protocol Systeming {
     /// - Throws: An error if the command fails.
     func async(_ arguments: [String]) throws
 
-    /// Returns the Swift version.
-    ///
-    /// - Returns: Swift version.
-    /// - Throws: An error if Swift is not installed or it exists unsuccessfully.
-    func swiftVersion() throws -> String
-
-    /// Returns the Swift version, including the build number.
-    ///
-    /// - Returns: Swift version including the build number.
-    /// - Throws: An error if Swift is not installed or it exists unsuccessfully.
-    func swiftlangVersion() throws -> String
-
     /// Runs /usr/bin/which passing the given tool.
     ///
     /// - Parameter name: Tool whose path will be obtained using which.
